Added more unit tests to newToken
