+++ bridge/contracts/libraries/aToken/ATokenBurner.sol
@@ -5,12 +5,33 @@ import "contracts/interfaces/IAToken.sol";
 import "contracts/utils/ImmutableAuth.sol";
 import "@openzeppelin/contracts/access/AccessControl.sol";
 
-contract ATokenBurner is ImmutableAToken, AccessControl {
-    bytes32 public constant BURNER_ROLE = keccak256("BURNER_ROLE");
+contract ATokenBurner is ImmutableAToken {
+    address internal _admin;
+    address internal _burner;
 
-    constructor() ImmutableFactory(msg.sender) ImmutableAToken() {}
+    modifier onlyAdmin() {
+        require(msg.sender == _admin, "onlyAdmin");
+        _;
+    }
 
-    function burn(address to, uint256 amount) public onlyRole(BURNER_ROLE) {
+    modifier onlyBurner() {
+        require(msg.sender == _burner, "onlyBurner");
+        _;
+    }
+
+    constructor() ImmutableFactory(msg.sender) ImmutableAToken() {
+        _admin = msg.sender;
+    }
+
+    function setAdmin(address admin_) public onlyFactory {
+        _admin = admin_;
+    }
+
+    function setBurner(address burner_) public onlyAdmin {
+        _burner = burner_;
+    }
+
+    function burn(address to, uint256 amount) public onlyBurner {
         IAToken(_ATokenAddress()).externalBurn(to, amount);
     }
 }
